Comment enhiriting cargo
Despite the fact that rust is already there it seams that inheriting
cargo requires another stable rust

Oddcoder / # emerge --ask relibc

 * IMPORTANT: 5 news items need reading for repository 'gentoo'.
 * Use eselect news read to view new items.

These are the packages that would be merged, in order:

Calculating dependencies... done!
[ebuild  N     ] virtual/rust-1.41.0  ABI_X86="(64) -32 (-x32)"
[ebuild  N    ~] sys-libs/relibc-9999  USE="-debug"

Would you like to merge these packages? [Yes/No] y^CInterrupted.

so I am disabling that for the sake of making the build process faster
and hoping that my guess is correct